requirements.txt
aiogram
asyncio
pillow
dotenv
git+https://github.com/SolovovChann/demotivator.git
